Convert session schedulers to actors
MSP-12110

Replace "SessionScheduler-<N>" threads with a pool of
Msf::SessionManager::Initializers registered as
:msf_session_manager_initializer_pool.  `Msf::SessionManager#schedule`
is replace with
`Celluloid::Actor[:msf_session_manager_initializer_pool].async.start_session`,
which takes a hash of arguments instead of a block.  A block is not
supported as blocks are not transmittable using DCell if we eventually
move to it.